# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AutoSEO API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new SEO project from a website URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/analyze:
    post:
      operationId: analyzeWebsite
      tags: [projects]
      summary: Analyze a website and extract business context using AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebsiteAnalysis"

  # ── Keywords ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        websiteUrl:
          type: string
        description:
          type: string
          nullable: true
        niche:
          type: string
          nullable: true
        targetAudience:
          type: string
          nullable: true
        language:
          type: string
        status:
          type: string
          enum: [pending, analyzing, active, paused]
        wordpressUrl:
          type: string
          nullable: true
        wordpressUsername:
          type: string
          nullable: true
        wordpressConnected:
          type: boolean
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, websiteUrl, language, status, createdAt, updatedAt]
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        websiteUrl:
          type: string
        language:
          type: string
          default: en
      required: [name, websiteUrl]
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        niche:
          type: string
        targetAudience:
          type: string
        language:
          type: string
        status:
          type: string
          enum: [pending, analyzing, active, paused]
    WebsiteAnalysis:
      type: object
      properties:
        niche:
          type: string
        targetAudience:
          type: string
        description:
          type: string
        suggestedKeywords:
          type: array
          items:
            type: string
      required: [niche, targetAudience, description, suggestedKeywords]
...