# lib/api-zod/src/generated/types/websiteAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoSEO API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WebsiteAnalysis {
  niche: string;
  targetAudience: string;
  description: string;
  suggestedKeywords: string[];
}